+++ dotnet/Models/CreateCategoryV2Request.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SheetsCatalogImport.Models
+{
+    public class CreateCategoryV2Request
+    {
+        [JsonProperty("parentId")]
+        public string ParentId { get; set; }
+
+        [JsonProperty("Name")]
+        public string Name { get; set; }
+    }
+}