+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Create catalog v2 brands & categories
+
 ## [0.2.1] - 2021-11-15
 
 ### Changed